requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.13
django-bootstrap-v5==1.0.11
django-filter==24.2
et-xmlfile==1.1.0
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==22.0.0
openpyxl==3.1.3
packaging==24.0
promise==2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
text-unidecode==1.3
ruff==0.4.5
